TurkeyMovement
Removed colliders from individual turkey components and replaced with a single sphere collider. Could have used a mesh collider but I didn't want to mess arround with generating a mesh, and the sphere collider seems to work fine.
The turkey now turns arround when it hits an object and seems to no longer go through walls.